pyproject.toml
[project]
name = "prodigy-task4"
version = "0.1.0"
description = "Lightweight Pix2Pix Inference"
requires-python = ">=3.12"
dependencies = [
    "onnxruntime==1.17.1",
    "numpy==1.26.4",
    "Pillow==10.2.0",
    "flask==3.0.2",
]

[tool.uv]
dev-dependencies = []